LayoutTests/webgpu/webgpu-basics-expected.txt

PASS [object WebGPU] is defined.
All tests complete.
PASS successfullyParsed is true

TEST COMPLETE
